Update progress to double

+++ Sources/Factory.swift
@@ -5,13 +5,13 @@ public final class Factory {
     public let resume = PassthroughSubject<Void, Never>()
     public let fail = PassthroughSubject<Void, Never>()
     public let finished = PassthroughSubject<Void, Never>()
-    public let progress = CurrentValueSubject<_, Never>(Float())
+    public let progress = CurrentValueSubject<_, Never>(Double())
     public let map: Map
     private weak var shooter: MKMapSnapshotter?
     private var shots: [Shot]
     private var result = [UInt8 : [UInt32 : [UInt32 : Data]]]()
     private var subs = Set<AnyCancellable>()
-    private let total: Float
+    private let total: Double
     private let points: [MKPointAnnotation]
     private let route: [MKRoute]
     private let timer = DispatchSource.makeTimerSource()
